core: Move Hangul decomposition into unicode.rs
